# products-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the products service.",
        "title": "Products Service API",
        "contact": {
            "name": "Subhankar Das",
            "url": "https://github.com/subhankardas",
            "email": "subhankardas831@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/products": {
            "get": {
                "description": "Returns list of all the products in store.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get list of products.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.Product"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Adds a new product to the store.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Add new product.",
                "parameters": [
                    {
                        "description": "Product Details",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "put": {
                "description": "Update product details to the store.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Update existing product.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Product Details",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "data.Product": {
            "type": "object",
            "required": [
                "name",
                "sku"
            ],
            "properties": {
                "description": {
                    "description": "Description",
                    "type": "string"
                },
                "id": {
                    "description": "Product ID",
                    "type": "integer"
                },
                "name": {
                    "description": "Product name",
                    "type": "string"
                },
                "price": {
                    "description": "Selling price",
                    "type": "number"
                },
                "sku": {
                    "description": "Stock ID, SKU has custom validator",
                    "type": "string"
                }
            }
        }
    }
}